refactor: 클린아키텍처 구조로 정리
