+++ notification/src/infrastructure/dynamo/provider.ts
@@ -1,0 +1,9 @@
+import * as dynamoose from "dynamoose";
+
+export default {
+    provide: "DYNAMOOSE",
+    useFactory: () => {
+        dynamoose.aws.ddb.local("http://localhost:8000"); // 로컬 설정
+        return dynamoose;
+    },
+};
