--- notification/src/infrastructure/ormEntity.ts
@@ -1,15 +1,0 @@
-import { Column, Entity, PrimaryColumn } from "typeorm";
-
-import { NotificationStatus } from "application/dto";
-
-@Entity("Notification")
-export default class NotificationORMEntity {
-    @PrimaryColumn()
-    event_id: string;
-
-    @Column({ type: "datetime" })
-    send_at: Date;
-
-    @Column({ type: "enum", enum: NotificationStatus, default: NotificationStatus.Pending })
-    status: NotificationStatus;
-}
