+++ notification/src/module.ts
@@ -3,18 +3,12 @@ import { ConfigModule } from "@nestjs/config";
 
 import NotificationService from "./application/service";
 
-import NotificationHttpController from "./presentation/controllers/httpController";
-import NotificationsMessageController from "./presentation/controllers/messageController";
-import DynamoRepository from "infrastructure/dynamoRepository";
-import * as dynamoose from "dynamoose";
+import NotificationHttpController from "presentation/controllers/httpController";
+import NotificationsMessageController from "presentation/controllers/messageController";
 
-const DynamooseProvider = {
-    provide: "DYNAMOOSE",
-    useFactory: () => {
-        dynamoose.aws.ddb.local("http://localhost:8000"); // 로컬 설정
-        return dynamoose;
-    },
-};
+import DynamooseProvider from "infrastructure/dynamo/provider";
+import DynamoRepository from "infrastructure/dynamo/repository";
+import DynamooseModel from "infrastructure/dynamo/model";
 
 @Module({
     providers: [
@@ -24,6 +18,7 @@ const DynamooseProvider = {
             provide: "INotificationRepository", // 인터페이스 제공
             useClass: DynamoRepository, // 구현체 연결
         },
+        DynamooseModel,
     ],
     controllers: [NotificationHttpController, NotificationsMessageController],
     imports: [
